refact: Convert to operate on a generic stream & an autobase
Previously I wanted to target replicating all cores on a corestore
assuming namespacing would be sufficient to limit the replication scope.
But since these cores need to be specified as inputs & outputs in my use
case, it was better to focus on syncing and adding cores on the
autobase.

The `get` argument allows for any core storage / management in the
application and the generic stream allows for any replication /
transport configuration (as long as its a stream). This support now more
use cases than just corestores.

All outputs and inputs are announced at creation with the option to
re-announce however often the application wants to via `announce()`. In
most cases announcing on creation is sufficient. Changes to existing
peers can be propagated by setting a timeout to announce at regular
intervals or via a more intelligent pattern. Currently there is not a
method to detect when new inputs or outputs are added to an autobase
without setting a interval that then reads the autobase's inputs and
outputs properties and sees if they've changed.

Finally a flexible filter for adding a core was added that provides:
- the key for the core to be added
- whether it is to be added as an output or input
- the session it came from
With this cores can be blacklisted.